refactors api_login.py to demonstrate ApiKey not working

+++ createDB.sql
@@ -3,12 +3,34 @@
  * The database has to be created first using the template in poorMansDbTemplate.zip 
  */
 
-/* Delete tables if they exist before creating them */
-DROP TABLE usersTb;
-DROP TABLE departmentTb;
-DROP TABLE departmentHierarchyTb;
-DROP TABLE professionTb;
-DROP TABLE attendanceTb;
+/* Delete foreign keys relations and tables if they exist before creating them */
+DECLARE @ConstraintName nvarchar(200)
+DECLARE @TableName nvarchar(200)
+DECLARE @SqlCommand nvarchar(1000)
+DECLARE cursor_constraints CURSOR FOR 
+    SELECT 
+        fk.name AS FK_name, 
+        t.name AS table_name
+    FROM 
+        sys.foreign_keys AS fk
+    INNER JOIN 
+        sys.tables AS t ON fk.parent_object_id = t.object_id
+OPEN cursor_constraints
+FETCH NEXT FROM cursor_constraints INTO @ConstraintName, @TableName
+WHILE @@FETCH_STATUS = 0
+BEGIN
+    SET @SqlCommand = 'ALTER TABLE ' + @TableName + ' DROP CONSTRAINT ' + @ConstraintName
+    EXEC sp_executesql @SqlCommand
+    FETCH NEXT FROM cursor_constraints INTO @ConstraintName, @TableName
+END
+CLOSE cursor_constraints
+DEALLOCATE cursor_constraints
+
+IF EXISTS (SELECT * FROM sys.tables WHERE name = N'usersTb') DROP TABLE usersTb;
+IF EXISTS (SELECT * FROM sys.tables WHERE name = N'departmentTb') DROP TABLE departmentTb;
+IF EXISTS (SELECT * FROM sys.tables WHERE name = N'departmentHierarchyTb') DROP TABLE departmentHierarchyTb;
+IF EXISTS (SELECT * FROM sys.tables WHERE name = N'professionTb') DROP TABLE professionTb;
+IF EXISTS (SELECT * FROM sys.tables WHERE name = N'attendanceTb') DROP TABLE attendanceTb;
 
 /* Create tables */
 CREATE TABLE usersTb (
@@ -20,18 +42,18 @@ CREATE TABLE usersTb (
     titles NVARCHAR(255),
     position_type NVARCHAR(255),
     profession_id UNIQUEIDENTIFIER NOT NULL,
-    department_id UNIQUEIDENTIFIER NOT NULL,
+    department_id NVARCHAR(255) NOT NULL,
     created_at DATETIMEOFFSET NOT NULL DEFAULT SYSDATETIMEOFFSET(),
     updated_at DATETIMEOFFSET NOT NULL DEFAULT SYSDATETIMEOFFSET(),
 );
 CREATE TABLE departmentTb (
-    department_id UNIQUEIDENTIFIER NOT NULL PRIMARY KEY DEFAULT NEWID(),
-    department_en NVARCHAR(255),
-    department_cz NVARCHAR(255)
+    department_id NVARCHAR(255) PRIMARY KEY,
+    department_en NVARCHAR(255) NOT NULL,
+    department_cz NVARCHAR(255) NOT NULL
 );
 CREATE TABLE departmentHierarchyTb (
-    department_id UNIQUEIDENTIFIER NOT NULL,
-    sub_department_id UNIQUEIDENTIFIER NOT NULL
+    department_id NVARCHAR(255) NOT NULL,
+    sub_department_id NVARCHAR(255) NOT NULL
 );
 CREATE TABLE professionTb (
     profession_id UNIQUEIDENTIFIER NOT NULL PRIMARY KEY DEFAULT NEWID(),
